+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/client/ProxyFileReadingFilter.java
@@ -10,41 +10,43 @@ import de.gematik.rbellogger.converter.RbelJexlExecutor;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.util.RbelMessagePostProcessor;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
 
 @RequiredArgsConstructor
+@Slf4j
 public class ProxyFileReadingFilter implements RbelMessagePostProcessor {
 
     private final RbelJexlExecutor jexlExecutor = new RbelJexlExecutor();
     private final String filterExpression;
     private final Map<String, RbelElement> deletedMessages = new HashMap<>();
-    private final Set<String> addedMessages = new HashSet<>();
 
     @Override
     public void performMessagePostConversionProcessing(RbelElement message, RbelConverter converter,
         JSONObject messageObject) {
-        final boolean keepMessage = jexlExecutor.matchesAsJexlExpression(message, filterExpression);
-        if (!keepMessage) {
-            if (!messageObject.has(PAIRED_MESSAGE_UUID)
-                || !addedMessages.contains(messageObject.getString(PAIRED_MESSAGE_UUID))) {
-                    deletedMessages.put(message.getUuid(), message);
-                    converter.removeMessage(message);
-                }
-        } else {
+        if (isKeepMessage(message)) {
             if (messageObject.has(PAIRED_MESSAGE_UUID)) {
                 RbelElement deletedPartner = deletedMessages.get(messageObject.getString(PAIRED_MESSAGE_UUID));
                 if (deletedPartner != null) {
+                    deletedMessages.remove(deletedPartner.getUuid());
+                }
+            }
+        } else {
+            if (messageObject.has(PAIRED_MESSAGE_UUID)) {
+                final var partnerMessage = deletedMessages.get(messageObject.getString(PAIRED_MESSAGE_UUID));
+                if (partnerMessage != null) {
                     converter.removeMessage(message);
-                    converter.addMessage(deletedPartner);
-                    converter.addMessage(message);
+                    converter.removeMessage(partnerMessage);
                 }
             } else {
-                addedMessages.add(message.getUuid());
+                deletedMessages.put(message.getUuid(), message);
             }
         }
+    }
+
+    private boolean isKeepMessage(RbelElement message) {
+        return jexlExecutor.matchesAsJexlExpression(message, filterExpression);
     }
 }